requirements.txt
azure-storage==0.32.0
Flask==0.11.1